usersテーブルに対するアクセスメソッドをModel_Userクラスから分離
