--- fuel/app/classes/repository/modelAccess.php
@@ -1,5 +1,0 @@
-<?php
-
-class Repository_ModelAccess
-{
-}
