+++ fuel/app/tests/view/view.php
@@ -7,7 +7,7 @@ class Repository_ViewTest extends PHPUnit_Framework_TestCase
 {
     public function testsignupのviewをuserオブジェクトのパラメータ付きで作成する()
     {
-        $user = Model_User::createUserObject();
+        $user = Repository_ModelUserAccess::createUserObject();
         $user->screen_name = 'userID';
         $user->name = 'userNAME';
         $user->password = 'userPASS';
@@ -19,7 +19,7 @@ class Repository_ViewTest extends PHPUnit_Framework_TestCase
 
     public function testloginのviewをuserオブジェクトのパラメータ付きで作成する()
     {
-        $user = Model_User::createUserObject();
+        $user = Repository_ModelUserAccess::createUserObject();
         $user->screen_name = 'userID';
         $user->password = 'userPASS';
         $expect = Repository_View::createLoginViewWithUserParams($user);
@@ -29,7 +29,7 @@ class Repository_ViewTest extends PHPUnit_Framework_TestCase
 
     public function testviewのオブジェクトに新たなパラメータ追加する()
     {
-        $view = Repository_View::createLoginViewWithUserParams(Model_User::createUserObject());
+        $view = Repository_View::createLoginViewWithUserParams(Repository_ModelUserAccess::createUserObject());
         $expect = Repository_View::setNewParam($view,'aho','boke');
         $this->assertEquals($expect->aho,'boke');
     }
